Preserves bucket_count to be the number of sharesets
Instead of changing storage_server.total_bucket_count to be the number
of shares, preserve the value for previous releases. We now use leasedb to
calculate this value instead of the crawler.

Adds storage_server.total_leased_sharecount and storage_server.total_leased_used_space
as additional parameters for stats.